requirements.txt
setuptools==57.5.0
pygame==2.1.2
pybullet-svl==3.1.6.4
imagecodecs==2022.9.26
numpy==1.23.3
numba==0.56.4
scipy==1.9.1
cffi==1.15.1
ipykernel==6.16
matplotlib==3.6.1
zarr==2.12.0
numcodecs==0.10.2
h5py==3.7.0
hydra-core==1.2.0
einops==0.4.1
tqdm==4.64.1
dill==0.3.5.1
scikit-video==1.1.11
scikit-image==0.19.3
pymunk==6.2.1
wandb==0.13.3
threadpoolctl==3.1.0
shapely==1.8.4
cython==0.29.32
imageio==2.22.0
imageio-ffmpeg==0.4.7
termcolor==2.0.1
tensorboard==2.10.1
tensorboardx==2.5.1
psutil==5.9.2
click==8.0.4
boto3==1.24.96
accelerate==0.13.2
datasets==2.6.1
diffusers==0.11.1
av==10.0.0
cmake==3.24.3
pillow==9.2.0
einops==0.4.1
transformers==4.45.2
absl-py==2.1.0
beautifulsoup4==4.12.3
bleach==6.1.0
defusedxml==0.7.1
jinja2==3.1.4
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
mistune==2.0.5
nbclient==0.10.0
nbformat==5.10.4
pandocfilters==1.5.1
tinycss2==1.3.0
traitlets==5.14.3
opencv-python==4.10.0.84
huggingface-hub==0.25.2
hydra-colorlog==1.2.0
hydra-core==1.2.0
omegaconf==2.3.0
torchvision==0.14.1
tokenizers==0.20.1
gym==0.26.2
